requirements.txt
python-telegram-bot==20.6
httpx==0.26.2